+++ NEWS.md
@@ -1,5 +1,10 @@
 # washi (development version)
 
+* Fonts are made available with {ragg} and {systemfonts} instead of {extrafont}. 
+
+* In `washi_flextable()` and `washi_theme()`, the default font is set to "sans"
+  if the provided font is not found in `systemfonts::system_fonts()$family`.
+
 # washi 0.1.0
 
 * Prepare for CRAN submission.